requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
PyYAML
Werkzeug